requirements.txt
django
django-extensions
psycopg2
pytest
pytest-django
factory-boy
